update environments as pdm
