+++ src/pdm_conda/environments/__init__.py
@@ -1,0 +1,4 @@
+from pdm_conda.environments.conda import CondaEnvironment
+from pdm.environments.base import BaseEnvironment
+
+__all__ = ["CondaEnvironment", "BaseEnvironment"]
